Revert "Rename channel pages/components"
This reverts commit 4340e4714f29d5fe7ed8d800662a02cf1c86988a.
